Adjust interfaces
- Use `f64` instead of `usize`
	- Use `Bar<f64, f64>`
	- Use `Pie<f64>`

+++ src/preprocessor/handlers/code_handler/bar_parser.rs
@@ -3,9 +3,9 @@ use crate::translate;
 use anyhow::Result;
 use plotly::Bar;
 
-pub fn parse_bar_data(bar_obj: &mut serde_json::Value) -> Result<Box<Bar<usize, usize>>> {
-    let x: Vec<usize> = must_translate(bar_obj, "x")?;
-    let y: Vec<usize> = must_translate(bar_obj, "y")?;
+pub fn parse_bar_data(bar_obj: &mut serde_json::Value) -> Result<Box<Bar<f64, f64>>> {
+    let x: Vec<f64> = must_translate(bar_obj, "x")?;
+    let y: Vec<f64> = must_translate(bar_obj, "y")?;
     let bar = Bar::new(x, y);
     let bar = translate! {
         bar,
